aptly_publish: accept alternative passphrase
If we did something like replace GPG key at runtime with a new one, it's
difficult (at least I don't know of a way) to tell the the aptly recipe
about the password.

That is, `node.override` can only be used at compile time. If we get our
GPG password at converge time (say, we load it from AWS SSM via aws
cookbook), we can not set the attribute with the right value.

As a work-around, we can allow specifying passphrase directly to the
resource, which we are allowed to populate with runtime-derived values.